+++ src/routes/piano/+page.svelte
@@ -5,7 +5,7 @@
 	import {get_audio_ctx} from '$lib/audio/audio_ctx';
 	import MidiInput from '$lib/audio/MidiInput.svelte';
 	import {MIDI_MAX, MIDI_MIN, type Midi} from '$lib/music/midi';
-	import {start_playing_note, type StopPlaying} from '$lib/audio/play_note';
+	import {start_playing, stop_playing} from '$lib/audio/play_note';
 	import InitMidiButton from '$lib/music/InitMidiButton.svelte';
 
 	const audio_ctx = get_audio_ctx();
@@ -17,20 +17,6 @@
 	// emphasize middle C to make it easier to orient oneself on a MIDI keyboard
 	const emphasized_keys = new Set([60 as Midi]);
 
-	// TODO allow playing notes at different volumes using velocity
-	const playing: Map<Midi, StopPlaying> = new Map();
-	const start_playing = (note: Midi): void => {
-		const current = playing.get(note);
-		if (current) return;
-		playing.set(note, start_playing_note(audio_ctx, note));
-	};
-	const stop_playing = (note: Midi) => {
-		const stop_playing_note = playing.get(note);
-		if (!stop_playing_note) return;
-		stop_playing_note?.();
-		playing.delete(note);
-	};
-
 	let midi_min: Midi = 48;
 	let midi_max: Midi = 96;
 
@@ -40,7 +26,7 @@
 <main bind:clientWidth>
 	<MidiInput
 		bind:this={midi_input}
-		on:note_start={(e) => start_playing(e.detail.note)}
+		on:note_start={(e) => start_playing(audio_ctx, e.detail.note)}
 		on:note_stop={(e) => stop_playing(e.detail.note)}
 	/>
 	<div class="piano-wrapper" style:padding="{piano_padding}px">
@@ -49,7 +35,7 @@
 				width={clientWidth - piano_padding * 2}
 				{midi_min}
 				{midi_max}
-				on:press={(e) => start_playing(e.detail)}
+				on:press={(e) => start_playing(audio_ctx, e.detail)}
 				on:release={(e) => stop_playing(e.detail)}
 				{emphasized_keys}
 			/>
